requirements.txt
# SnapAI Dependencies
# Core AI and ML
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# GUI Framework
PyQt5>=5.15.0

# Web Framework
Flask>=2.3.0
websockets>=11.0.0

# Image Processing
Pillow>=9.5.0
numpy>=1.24.0

# System Integration
keyboard>=0.13.5
psutil>=5.9.0

# Development and Monitoring
watchdog>=3.0.0

# HTTP Requests (for health checks)
requests>=2.31.0
